refactor: extract ArenaContract into separate modules (state, players, rounds, payout)
