+++ contract/arena/src/commit_reveal_tests.rs
@@ -150,7 +150,7 @@ fn test_reveal_after_deadline() {
     let commitment = create_commitment(&env, &player1, Choice::Heads, &nonce);
     client.commit_choice(&player1, &1, &commitment);
     
-    set_ledger_sequence(&env, round.reveal_deadline_ledger + 1);
+    set_ledger_sequence(&env, round.round_deadline_ledger + 11); // past reveal deadline (commit_deadline + round_speed)
     client.reveal_choice(&player1, &1, &Choice::Heads, &nonce);
 }
