+++ contract/arena/src/errors.rs
@@ -1,0 +1,48 @@
+use soroban_sdk::contracterror;
+
+#[contracterror]
+#[derive(Copy, Clone, Debug, Eq, PartialEq, PartialOrd, Ord)]
+#[repr(u32)]
+pub enum ArenaError {
+    AlreadyInitialized = 1,
+    InvalidRoundSpeed = 2,
+    RoundAlreadyActive = 3,
+    NoActiveRound = 4,
+    SubmissionWindowClosed = 5,
+    SubmissionAlreadyExists = 6,
+    RoundStillOpen = 7,
+    RoundDeadlineOverflow = 8,
+    NotInitialized = 9,
+    Paused = 10,
+    ArenaFull = 11,
+    AlreadyJoined = 12,
+    InvalidAmount = 13,
+    NoPrizeToClaim = 14,
+    AlreadyClaimed = 15,
+    ReentrancyGuard = 16,
+    NotASurvivor = 17,
+    GameAlreadyFinished = 18,
+    TokenNotSet = 19,
+    MaxSubmissionsPerRound = 20,
+    PlayerEliminated = 21,
+    WrongRoundNumber = 22,
+    NotEnoughPlayers = 23,
+    InvalidCapacity = 24,
+    NoPendingUpgrade = 25,
+    TimelockNotExpired = 26,
+    GameNotFinished = 27,
+    TokenConfigurationLocked = 28,
+    UpgradeAlreadyPending = 29,
+    WinnerAlreadySet = 30,
+    WinnerNotSet = 31,
+    AlreadyCancelled = 32,
+    InvalidMaxRounds = 33,
+    NameTooLong = 34,
+    NameEmpty = 35,
+    DescriptionTooLong = 36,
+    NoCommitment = 37,
+    CommitmentMismatch = 38,
+    RevealDeadlinePassed = 39,
+    CommitDeadlinePassed = 40,
+    AlreadyCommitted = 41,
+}
